requirements.txt
streamlit
langchain
langchain-community
langchain-groq
pandas
plotly
sqlalchemy
python-dotenv
tabulate